refactor(pagination): rename to consistent naming convention
Rename CustomPagination to Pagination and TinyPagination to
PaginationTiny for consistency with component naming standards.
