+++ lib/components/Pagination/PaginationTiny.vue
